requirements.txt
openai
langchain
streamlit
streamlit-chat
tiktoken
python-dotenvpip
langchain-community